add Sol, Star, HeliocentricOrbitalBody, Earth classes, refactor kepler calc date param and earth.position.date object
